requirements.txt
fastapi==0.95.2
pydantic==1.10.14
uvicorn==0.29.0
python-dotenv==1.0.1